extracting validation
